fix: incorrect 500 in static pages
