+++ app/(header-default)/faq/[slug]/page.tsx
@@ -1,14 +1,27 @@
 import { Metadata } from 'next';
+import { notFound } from 'next/navigation';
 import Breadcrumb from '#components-ui/breadcrumb';
 import ButtonLink from '#components-ui/button';
 import TextWrapper from '#components-ui/text-wrapper';
 import { RenderMarkdownServerOnly } from '#components/markdown';
 import { allFaqArticles, getFaqArticle } from '#models/article/faq';
+import { Exception } from '#models/exceptions';
+import { logWarningInSentry } from '#utils/sentry';
 import { AppRouterProps } from '#utils/server-side-helper/app/extract-params';
-import { redirectFAQPageNotFound } from '#utils/server-side-helper/app/redirect-faq-not-found';
 
 type IParams = {
   slug: string;
+};
+
+// should not happen since we declared generateStaticParams
+const redirectFAQPageNotFound = (slug: string) => {
+  logWarningInSentry(
+    new Exception({
+      name: 'FAQPageNotFound',
+      context: { slug },
+    })
+  );
+  notFound();
 };
 
 export default async function FAQArticle(props: AppRouterProps) {
