+++ app/(header-default)/faq/modifier/[slug]/page.tsx
@@ -1,4 +1,5 @@
 import { Metadata } from 'next';
+import { notFound } from 'next/navigation';
 import Breadcrumb from '#components-ui/breadcrumb';
 import ButtonLink from '#components-ui/button';
 import TextWrapper from '#components-ui/text-wrapper';
@@ -6,11 +7,23 @@ import {
   allDataToModify,
   getDataToModify,
 } from '#models/administrations/data-to-modify';
+import { Exception } from '#models/exceptions';
+import { logWarningInSentry } from '#utils/sentry';
 import { AppRouterProps } from '#utils/server-side-helper/app/extract-params';
-import { redirectFAQPageNotFound } from '#utils/server-side-helper/app/redirect-faq-not-found';
 
 type IParams = {
   slug: string;
+};
+
+// should not happen since we declared generateStaticParams
+const redirectFAQPageNotFound = (slug: string) => {
+  logWarningInSentry(
+    new Exception({
+      name: 'FAQPageNotFound',
+      context: { slug },
+    })
+  );
+  notFound();
 };
 
 export default async function FAQArticle(props: AppRouterProps) {
