--- utils/server-side-helper/app/redirect-faq-not-found.ts
@@ -1,13 +1,0 @@
-import { redirect } from 'next/navigation';
-import { Exception } from '#models/exceptions';
-import { logWarningInSentry } from '#utils/sentry';
-
-export const redirectFAQPageNotFound = (slug: string) => {
-  logWarningInSentry(
-    new Exception({
-      name: 'FAQPageNotFound',
-      context: { slug },
-    })
-  );
-  redirect('/404');
-};
